Move data generator instructions before walkthrough
There's a necessary one-time setup step to create some fake data for
SKUs and categories. It was introduced in the "Model" section, but
that comes after the walkthrough. The walkthrough fails without the
data, so this would be a stumbling block for new users.

This change moves the instructions earlier. It also renames the
generator from repl.clj to the more-explicit create-data.clj.